refactor: 댓글방 상태 관련 API 엔드포인트 수정 및 패키지 변경
