+++ backend/src/main/java/com/zzang/chongdae/comment/controller/CommentController.java
@@ -3,7 +3,6 @@ package com.zzang.chongdae.comment.controller;
 import com.zzang.chongdae.comment.service.CommentService;
 import com.zzang.chongdae.comment.service.dto.CommentAllResponse;
 import com.zzang.chongdae.comment.service.dto.CommentRoomAllResponse;
-import com.zzang.chongdae.comment.service.dto.CommentRoomStatusResponse;
 import com.zzang.chongdae.comment.service.dto.CommentSaveRequest;
 import jakarta.validation.Valid;
 import java.net.URI;
@@ -12,7 +11,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -42,14 +40,6 @@ public class CommentController {
             @PathVariable(value = "offering-id") Long offeringId,
             @RequestParam(value = "member-id") Long loginMemberId) {
         CommentAllResponse response = commentService.getAllComment(offeringId, loginMemberId);
-        return ResponseEntity.ok(response);
-    }
-
-    @PutMapping("/comments/{offering-id}/status")
-    public ResponseEntity<CommentRoomStatusResponse> updateCommentRoomStatus(
-            @PathVariable(value = "offering-id") Long offeringId,
-            @RequestParam(value = "member-id") Long loginMemberId) {
-        CommentRoomStatusResponse response = commentService.updateCommentRoomStatus(offeringId, loginMemberId);
         return ResponseEntity.ok(response);
     }
 }
